Update CSS path, Refactor Home component
