+++ src/views/Home.js
@@ -1,5 +1,6 @@
+import { Link } from 'react-router-dom';
 import UserDashboard from '../components/UserDashboard';
-import '../assets/Home.css';
+import '../assets/stylesheets/Home.css';
 
 const Home = () => {
   const user = JSON.parse(localStorage.getItem('user'))
@@ -8,11 +9,29 @@ const Home = () => {
 
   return (
     <div className="Home">
-      <h1>
-        Welcome!
-        { user && user }
-      </h1>
-      <UserDashboard />
+      {
+        user
+          ? (
+            <>
+              <h1>
+                Welcome!
+                { user && user }
+              </h1>
+              <UserDashboard />
+            </>
+          )
+          : (
+            <>
+              <h1>Wellcome!</h1>
+              <p>
+                Storm Mountain is a winter sports resort.
+                To book a lesson please sign up and sign up
+                and visit the activities page.
+              </p>
+              <Link to="/signin" className="Button">Sign In</Link>
+            </>
+          )
+      }
     </div>
   );
 };
